+++ src/mbgl/actor/scheduler.cpp
@@ -28,7 +28,7 @@ std::shared_ptr<Scheduler> Scheduler::GetBackground() {
     std::shared_ptr<Scheduler> scheduler = weak.lock();
 
     if (!scheduler) {
-        weak = scheduler = std::make_shared<ThreadPool>(4);
+        weak = scheduler = std::make_shared<ThreadPool>();
     }
 
     return scheduler;